length of string

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -6,9 +6,9 @@
  *
  * Return: The length of the string - @str.
  */
-size_t _strlen(char *str)
+int _strlen(char *str)
 {
-	size_t length = 0;
+	int length = 0;
 
 	while (*str++)
 		length++;
